groundwork
Excited about this project.

I plan on creating a site with a homepage that displays icons for every league of legends champion.

When you click an icon, it will route you to the champion page where you can see an AI generated short story about that champion, as well as some general information on the champion.