champions page wrapping up
makes an API call to gpt3 to generate a personalized short story incorporating the users name from an input field.

Pretty much wrapping up the champion page template so all I need to do for the front end is recreate the page & story for the 5 other champions, then we can shift to the backend and wrap up the project.